pyproject.toml
[project]
name = "agent-operator-lab"
version = "0.1.0"
description = "Hands-on labs for operating Azure-based AI agents."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  # Agent Framework (MAF) & Extensions
  "agent-framework==1.0.0b251223",
  "agent-framework-azure-ai==1.0.0b251223",
  "agent-framework-core==1.0.0b251223",
  "mcp>=1.24.0",
  # Azure Core Packages
  "azure-identity==1.25.1",
  "azure-core==1.37.0",
  "azure-mgmt-resource==24.0.0",

  # Azure AI Services
  "azure-ai-projects==2.0.0b2",
  "azure-ai-evaluation==1.8.0",
  "azure-ai-inference==1.0.0b8",
  "azure-search-documents==11.5.2",
  "azure-ai-agentserver-agentframework==1.0.0b7",

  # OpenAI
  "openai==2.13.0",
  
  # HTTP Client
  "httpx>=0.24.1",
  "requests>=2.31.0",

  # Utilities
  "python-dotenv==1.1.1",
  "tiktoken==0.8.0",
  "pyyaml~=6.0.2",

  # Azure Monitor & OpenTelemetry
  "azure-mgmt-loganalytics==13.1.1",
  "azure-monitor-opentelemetry==1.8.1",
  "azure-mgmt-applicationinsights==4.1.0",

  # Jupyter / Notebook Support
  "ipykernel>=6.29.0",
  "jupyter>=1.0.0",
  "ipywidgets==8.1.5",
]